+++ packages/api/lib/reconciliationReport.js
@@ -1,3 +1,5 @@
+//@ts-check
+
 'use strict';
 
 const isEqual = require('lodash/isEqual');
@@ -8,6 +10,14 @@ const { constructCollectionId, deconstructCollectionId } = require('@cumulus/mes
 const Logger = require('@cumulus/logger');
 
 const log = new Logger({ sender: '@api/lambdas/create-reconciliation-report' });
+
+/**
+ * @typedef {import('../lib/types').RecReportParams } RecReportParams
+ * @typedef {import('../lib/types').EnhancedNormalizedRecReportParams }
+ * EnhancedNormalizedRecReportParams
+ * @typedef {import('../lib/types').NormalizedRecReportParams } NormalizedRecReportParams
+ * @typedef {import('./reconciliationReport-types').ReportHeader } ReportHeader
+ */
 
 /**
  * Extra search params to add to the cmrGranules searchConceptQueue
@@ -23,6 +33,7 @@ function cmrGranuleSearchParams(recReportParams) {
   return [];
 }
 
+// TODO: remove
 /**
  * Prepare a list of collectionIds into an _id__in object
  *
@@ -36,7 +47,7 @@ function searchParamsForCollectionIdArray(collectionIds) {
 
 /**
  * @param {string} dateable - any input valid for a JS Date contstructor.
- * @returns {number} - primitive value of input date string or undefined, if
+ * @returns {number | undefined} - primitive value of input date string or undefined, if
  *                     input string not convertable.
  */
 function dateToValue(dateable) {
@@ -49,6 +60,7 @@ function dateStringToDateOrNull(dateable) {
   return !Number.isNaN(date.valueOf()) ? date : undefined;
 }
 
+// TODO - Remove this in 3806 PR #3
 /**
  *
  * @param {Object} params - request params to convert to Elasticsearch params
@@ -71,12 +83,12 @@ function convertToESCollectionSearchParams(params) {
  * convertToDBCollectionSearchObject      - Creates Postgres search object from
  *                                          InternalRecReport Parameters
  * @param {Object} params                 - request params to convert to database params
- * @param {[Object]} params.collectionIds - List containing single Collection object
+ * @param {string[]} [params.collectionIds] - List containing single Collection object
  *                                          multiple or no collections will result in a
  *                                          search object without a collection object
- * @param {moment} params.endTimestamp    - ending report datetime ISO Timestamp
- * @param {moment} params.startTimestamp  - beginning report datetime ISO timestamp
- * @returns {[Object]}                    - array of objects of desired
+ * @param {string} [params.endTimestamp]    - ending report datetime ISO Timestamp
+ * @param {string} [params.startTimestamp]  - beginning report datetime ISO timestamp
+ * @returns {Object[]}                    - array of objects of desired
  *                                          parameters formatted for database collection
  *                                          search
  */
@@ -121,8 +133,8 @@ function convertToESGranuleSearchParams(params) {
 /**
  * Convert reconciliation report parameters to PostgreSQL database search params.
  *
- * @param {Object} params - request params to convert to database params
- * @returns {Object} object of desired parameters formated for database granule search
+ * @param {EnhancedNormalizedRecReportParams} params - request params to convert to database params
+ * @returns object of desired parameters formatted for database granule search
  */
 function convertToDBGranuleSearchParams(params) {
   const {
@@ -152,7 +164,7 @@ function convertToDBGranuleSearchParams(params) {
  * convert to es search parameters using createdAt for report time range
  *
  * @param {Object} params - request params to convert to Elasticsearch params
- * @returns {Object} object of desired parameters formated for Elasticsearch.
+ * @returns {Object} object of desired parameters formatted for Elasticsearch.
  */
 function convertToESGranuleSearchParamsWithCreatedAtRange(params) {
   const searchParamsWithUpdatedAt = convertToESGranuleSearchParams(params);
@@ -167,7 +179,7 @@ function convertToESGranuleSearchParamsWithCreatedAtRange(params) {
 /**
  *
  * @param {Object} params - request params to convert to orca params
- * @returns {Object} object of desired parameters formated for orca
+ * @returns {Object} object of desired parameters formatted for orca
  */
 function convertToOrcaGranuleSearchParams(params) {
   const { collectionIds, granuleIds, providers, startTimestamp, endTimestamp } = params;
@@ -183,12 +195,8 @@ function convertToOrcaGranuleSearchParams(params) {
 /**
  * create initial report header
  *
- * @param {Object} recReportParams - params
- * @param {Object} recReportParams.reportType - the report type
- * @param {moment} recReportParams.createStartTime - when the report creation was begun
- * @param {moment} recReportParams.endTimestamp - ending report datetime ISO Timestamp
- * @param {moment} recReportParams.startTimestamp - beginning report datetime ISO timestamp
- * @returns {Object} report header
+ * @param {EnhancedNormalizedRecReportParams} recReportParams - params
+ * @returns {ReportHeader} report header
  */
 function initialReportHeader(recReportParams) {
   const {